+++ doc/changelog.rst
@@ -9,6 +9,8 @@ Version 0.8.4
   interruptions (IRQ).
 * ``get_dates()`` methods now return ``None`` rather than an empty tuple
   if runs don't have the ``date`` metadata.
+* The ``--stdout`` internal option has been removed, replaced by a new
+  ``--pipe`` option. Workers can now use stdout for regular messages.
 
 Version 0.8.3 (2016-11-03)
 --------------------------